Chunk 7 (final): both modules compile with 0 errors for MC 26.1.2
Neoforge module fixes:
- Delete conflicting Neo* bridge files (NeoParticles, NeoConsumeEffects,
  NeoBlockPredicateTypes, NeoMobEffects, NeoPotions) — resolved 84 static
  final errors
- Add JEI and terrablender as compile dependencies
- Fix critereon→criterion, GuiGraphics→GuiGraphicsExtractor in neoforge
- Fix BlockRenderDispatcher→BlockModelResolver, noCollission→noCollision
- Fix hasPermission, sendSystemMessage, displayClientMessage signatures
- Fix CookingBookCategory import, ItemStack→ItemLike in recipe builders
- Fix BiomeTagProvider: getOrCreateRawBuilder + addTag(Identifier)
- Fix RecipeProvider smelting/blasting signatures
- Fix AdvancementProvider display() method signature
- Fix particle registration API (registerSpecial/registerSpriteSet)
- Fix Block methods (getAnalogOutputSignal, getTicker)
- Fix JEI plugin stubs (JEI 21.x incompatible with MC 26.1.2 ResourceLocation→Identifier)
- Fix PlatformImpl: FMLEnvironment.dist→getDist(), production→isProduction()
- Fix PacketNetworkImpl: sendToServer varargs, ClientPacketDistributor
- Fix entity/enchantment/effect listener type casts
- Add access widener entries for neoforge model generation
- Various import and type inference fixes

+++ neoforge/src/main/java/grill24/potionsplus/event/EffectListeners.java
@@ -37,7 +37,7 @@ public class EffectListeners {
         if (event.getEntity() instanceof LivingEntity livingEntity) {
             BouncingEffect.onFall(livingEntity);
         }
-        if (BouncingEffect.onLivingFall(event.getEntity(), event.getDistance())) {
+        if (BouncingEffect.onLivingFall(event.getEntity(), (float) event.getDistance())) {
             event.setCanceled(true);
         }
     }
